refactor: add baseurl for back button and click outside functionality
